Move test types into respective test functions
To make it easier to add new types used for testing the crate's
functionality in the future, this change moves all such existing types
into the respective test functions. This way, we reduce the possibility
for name collisions or the need to get overly creative with names to
begin with, respectively.