Tag functions and methods as #[inline]
Unless a function (or method) is generic, Rust generally does not keep
sufficient information around to be able to inline calls across crate
boundaries. When it comes to small functions, that can be detrimental to
both speed and binary size.
With this change we tag more functions with #[inline] in an attempt to
give Rust more freedom when it comes to inlining function bodies in
consuming crates.